+++ LearnOpenGL/src/VertexBuffer.h
@@ -6,7 +6,7 @@ class VertexBuffer
 {
 public:
     VertexBuffer() : id(0) {}
-    VertexBuffer(const void* data, size_t size, GLenum mode);
+    VertexBuffer(const void* data, unsigned int size, GLenum mode);
     ~VertexBuffer() {}
 
     void Bind();
